requirements.txt
# autogen
autogen-agentchat~=0.2
python-dotenv